+++ test/ruby/test_process.rb
@@ -1451,10 +1451,10 @@ class TestProcess < Test::Unit::TestCase
       assert_equal(s, s)
       assert_equal(s, s.to_i)
 
-      assert_warn(/\bUse .*Process::Status/) do
+      assert_deprecated_warn(/\buse .*Process::Status/) do
         assert_equal(s.to_i & 0x55555555, s & 0x55555555)
       end
-      assert_warn(/\bUse .*Process::Status/) do
+      assert_deprecated_warn(/\buse .*Process::Status/) do
         assert_equal(s.to_i >> 1, s >> 1)
       end
       assert_raise(ArgumentError) do